Feat: Define DataSource Kind
The DataSource kind is another abstraction that utilised the Connection. The DataSource allows users to define additional metadata to fetch the DataSource or send data to the DataSource. It defines a schema field which allows users to define a Connection from which to fetch the schema for the DataSource. Additionally it has a metadata field for defining additional properties, e.g. the table name of where the data resides.

As with the Connection, there is a DataSourceType to validate the DataSource dynamically. This allows users to define the structure of the DataSource.

Signed-off-by: Lennard Eijsackers <lennardeijsackers92@gmail.com>