Feat/add DataSet kind (#10)
* Feat: Define DataSet Kind

The DataSet kind is another abstraction that utilised the Connection. The DataSet allows users to define additional metadata to fetch the DataSet or send data to the DataSet. It defines a schema field which allows users to define a Connection from which to fetch the schema for the DataSet. Additionally it has a metadata field for defining additional properties, e.g. the table name of where the data resides.

As with the Connection, there is a DataSetType to validate the DataSource dynamically. This allows users to define the structure of the DataSet.

Signed-off-by: Lennard Eijsackers <lennardeijsackers92@gmail.com>